# DGVFF_WS_CLIENT/swagger/docs/library/public-api/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a DGVFF Client Server.  You can find out more about DGVFF project at <a href=\"http://dgvff.com\">http://dgvff.com</a> or <a href=\"http://ideabytes.com\">http://ideabytes.com</a>. ",
    "version": "1.0.0",
    "title": "DGVFF_WS_CLIENT",
    "termsOfService": "http://dgvff.com/",
    "contact": {
      "url": "http://dgvff.com",
      "name": "Your pals at Swagger",
      "email": "info.dfvff.com"
    },
    "license": {
      "name": "localhost:80",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:80",
  "basePath": "/DGVFF_WS_CLIENT",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
        "/api/web/GetBillOfladdingDetails.json": {
      "post": {
        "tags": [
          "DGVFF_WS_CLIENT"
        ],
        "summary": "Get BillOfladding Details",
        "description": "Get BillOfladding Details",
        "operationId": "GetBillOfladdingDetails",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "BillOfladdingInfo",
            "description": "BillOfladding",
            "required": true
          }
        ],
        "responses": {
          "405": {
            "description": "Validation exception"
          },
          "404": {
            "description": "BillOfladding not found"
          },
          "400": {
            "description": "Invalid ID supplied"
          }
        },
        "security": [
          {
            "petstore_auth": [
              "write:DGVFF_WS_CLIENT",
              "read:DGVFF_WS_CLIENT"
            ]
          }
        ]
      }
    },
...
},
...
}